Basic splitting into components
