+++ WorkloadManagementSystem/Service/MatcherHandler.py
@@ -16,7 +16,8 @@ from DIRAC.FrameworkSystem.Client.MonitoringClient     import gMonitor
 
 from DIRAC.WorkloadManagementSystem.DB.JobDB           import JobDB
 from DIRAC.WorkloadManagementSystem.DB.TaskQueueDB     import TaskQueueDB
-from DIRAC.WorkloadManagementSystem.Client.Matcher     import Matcher, Limiter
+from DIRAC.WorkloadManagementSystem.Client.Matcher     import Matcher
+from DIRAC.WorkloadManagementSystem.Client.Limiter     import Limiter
 
 gJobDB = False
 gTaskQueueDB = False
@@ -58,8 +59,8 @@ def sendNumTaskQueues():
 class MatcherHandler( RequestHandler ):
 
   def initialize( self ):
+    self.matcher = Matcher( jobDB = gJobDB, tqDB = gTaskQueueDB )
     self.limiter = Limiter()
-    self.matcher = Matcher()
 
 ##############################################################################
   types_requestJob = [ [StringType, DictType] ]
@@ -68,10 +69,16 @@ class MatcherHandler( RequestHandler ):
         one matching the agent's site capacity
     """
 
-    result = self.matcher.selectJob( resourceDescription )
+    resourceDescription['Setup'] = self.serviceInfoDict['clientSetup']
+    credDict = self.getRemoteCredentials()
+
+    try:
+      result = self.matcher.selectJob( resourceDescription, credDict )
+    except RuntimeError, rte:
+      self.log.error( "Error requesting job: ", rte )
+      return S_ERROR( "Error requesting job" )
     gMonitor.addMark( "matchesDone" )
-    if result[ 'OK' ]:
-      gMonitor.addMark( "matchesOK" )
+    gMonitor.addMark( "matchesOK" )
     return result
 
 ##############################################################################
