--- WorkloadManagementSystem/Service/test/Test_Matcher.py
@@ -1,23 +1,0 @@
-ceDict = {'Architecture': 'x86_64-slc6',
- 'CEQueue': 'jenkins-queue_not_important',
- 'CPUNormalizationFactor': '9.5',
- 'CPUScalingFactor': '9.5',
- 'CPUTime': 1080000,
- 'CPUTimeLeft': 5000,
- 'DIRACVersion': 'v8r0p1',
- 'FileCatalog': 'LcgFileCatalogCombined',
- 'GridCE': 'jenkins.cern.ch',
- 'GridMiddleware': 'DIRAC',
- 'LHCbPlatform': 'x86_64-slc5-gcc43-opt',
- 'LocalSE': ['CERN-SWTEST'],
- 'MaxTotalJobs': 100,
- 'MaxWaitingJobs': 10,
- 'OutputURL': 'gsiftp://localhost',
- 'PilotBenchmark': 9.5,
- 'PilotReference': 'somePilotReference',
- 'Platform': 'x86_64-slc6',
- 'ReleaseProject': 'LHCb',
- 'ReleaseVersion': 'v8r0p1',
- 'Setup': 'LHCb-Certification',
- 'Site': 'DIRAC.Jenkins.ch',
- 'WaitingToRunningRatio': 0.05}
